Transfered the data files from the GSON test project

+++ app/src/main/java/nl/jeroenhd/app/bcbreader/ChapterReadingActivity.java
@@ -66,7 +66,7 @@ public class ChapterReadingActivity extends AppCompatActivity {
     void SetupData()
     {
         mPages = new ArrayList<>();
-        for (int i = 0; i < 20; i++)
+        for (Double i = 0.0; i < 20; i++)
         {
             mPages.add(new Page("Example commentary", i+1, 0.0));
         }
